+++ packages/tts-editor/src/command/saveAndPlay.ts
@@ -5,5 +5,5 @@ export const saveAndPlay = (adapter: TTSAdapter) => () => {
 };
 
 export const saveAndPlayBundled = (adapter: TTSAdapter) => () => {
-  adapter.saveAndPlay(true);
+  adapter.saveAndPlay("bundle");
 };