+++ packages/tts-editor/src/command/unbundleLibrary.ts
@@ -1,0 +1,5 @@
+import { TTSAdapter } from "../ttsAdapter";
+
+export default (adapter: TTSAdapter) => () => {
+  adapter.unbundleLibrary();
+};